Version 10.1 (04/11/2025) - Update to -g and -p parameters
The -g parameter accepts more than one partitioning at a time. The partitionings must be separated by a slash ("/").
The logfile shows all the partitionings that were specified.
If a method is to perform partitioning, but has not been specified in -p, it is run anyway, unless it has already been run previously.
⁠Running all methods with graphics is the default, if not specified with -g.
If only -g is specified, it will run only the partitionings and eventually the method, if it has not been run previously.
⁠If both the -p and -g parameters are specified, MPACT will execute the partitioning(s) specified in the -g parameter and the graphics of the methods specified in the -p parameter.
Correction of the description of the -g parameter in the help page.
In addition to being able to perform more than one partitioning at a time, it is now possible to generate more than one partitioning for the same method.